+++ NEWS.md
@@ -5,6 +5,9 @@
 
 * Fix an error with the validator of variable names in `silviculture::Inventory` S7 object.
 
+## Enhancements
+
+* `silv_volume()`: it assumed diameter to be in meters. Now the diameter must be given in centimeters. An informing message was added to the function.
 
 # silviculture 0.1.0